Fix stream revision # after snapshot
In getFromSnapshot if there are no events in the stream after the snapshot, the
returned stream object had a lastRevision of -1 which causes the next event to
be created with a revision number of 0, breaking the event stream. This fix
set's the last revision equal to the snapshot's revision number.