Early abort when events were fetched without limit
When using MongoDB as backend and calling `es.getEvents((err, events) => { ... })`, any call to `events.next` will result in an error, as the current logic adds the default `-1` limit to the `skip` value of `0`, resulting in an invalid negative skip value. Invoking the callback in this case with an empty array should be a graceful fix.