Fix dynamodb concurrency (#110)
Currently the condition expression that is supposed to protect against concurrency violations will be invoked when both commits are created using the exact same commit timestamp. This is because condition expressions are only checked against items with the same key. The fix for this is to remove the timestamp from the key, leaving the aggregateId, aggregate, context and streamRevision.