Fix concurrency issue in createTableIfNotExists (#118)
During concurrent operations there is no guarantee that a table is not
created between checking it's existance (using describeTable) and making
the call to create the table. Currently, when this happens, it is
erroring out with the 'Cannot create preexisting table' error.

Even ensuring that the init promise is resolved before calling it again
is not enough because the call to describeTable is eventually
consistent.

The fix is to swallow the error when the table already exists. At the
end of the day, the system is in the state that we want it to be anyway
(the table exists).